+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Fixes
 - craft objective when item on courser
+- NPC hider for not spawned NPCs 
 ### Security
 
 ## [1.12.2] - 2021-03-14